Changed getIni bool defaults from strings
Surely this was an underlying bug in my default handling? 

Yes, the correct value was being written to the INI file (and next time would be correctly with a 'getboolean' method) however this first call would return the 'False' value as a *string*, which an upstream test would interpret as True.