Added isUserLocal [incomplete]
Aborts ap and noap (*on bookworm*) if the connected user's session has an IP address, making it either a wired session OR wireless. (If wireless you'll be disconnected during the delete/add network steps, leaving you high and dry).

Only a local (direct-connect) to the Pi will succeed at this stage.